requirements.txt
alembic==0.9.9
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
Mako==1.0.7
MarkupSafe==1.0
nbformat==4.4.0
pkg-resources==0.0.0
plotly==2.5.1
pyserial==3.4
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
traitlets==4.3.2
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
